Move these tests to the right directory
